project/backend/requirements.txt
playwright
playwright-stealth
google-genai
pillow
requests
python-dotenv
psycopg[pool,binary]
fastmcp
pgvector==0.2.5
fastapi
openai
beautifulsoup4
httpx 
uvicorn[standard]